API changes to silence clang warnings

+++ src/include/compiler/Compiler.h
@@ -45,10 +45,10 @@ class Compiler {
   LogicalFactory _logicalfactory;
   MixtureFactory _mixfactory1;
   MixtureFactory _mixfactory2;
-  std::map<std::string, std::vector<int> > _node_array_bounds;
-  std::map<std::pair<std::string, Range>, std::set<int> > _umap;
+  std::map<std::string, std::vector<unsigned long> > _node_array_bounds;
+  std::map<std::pair<std::string, Range>, std::set<unsigned long> > _umap;
   std::set<std::string> _lhs_vars;
-  std::map<std::pair<std::vector<unsigned int>, std::vector<double> >,
+  std::map<std::pair<std::vector<unsigned long>, std::vector<double> >,
       ConstantNode *> _cnode_map;
   
   Node *getArraySubset(ParseTree const *t);
@@ -72,18 +72,18 @@ class Compiler {
   Node * constFromTable(ParseTree const *p);
   void addDevianceNode();
   Node *getConstant(double value, unsigned int nchain, bool observed);
-  Node *getConstant(std::vector<unsigned int> const &dim, 
+  Node *getConstant(std::vector<unsigned long> const &dim, 
 		    std::vector<double> const &value,
 		    unsigned int nchain, bool observed);
   void getLHSVars(ParseTree const *rel);
 public:
-  bool indexExpression(ParseTree const *t, std::vector<int> &value);
+  bool indexExpression(ParseTree const *t, std::vector<unsigned long> &value);
   BUGSModel &model() const;
   Node * getParameter(ParseTree const *t);
   /**
    * @param model Model to be created by the compiler.
    *
-   * @param datatab Data table, mapping a variable name onto a
+   * @param data_table Data table, mapping a variable name onto a
    * multi-dimensional array of values. This is required since some
    * constant expressions in the BUGS language may depend on data
    * values.
