+++ src/include/function/ArrayLogDensity.h
@@ -15,18 +15,18 @@ namespace jags {
 	ArrayDist const *_dist;
     public:
 	ArrayLogDensity(ArrayDist const *dist);
-	std::vector<unsigned int> dim(
-	    std::vector<std::vector<unsigned int> > const &dims,
+	std::vector<unsigned long> dim(
+	    std::vector<std::vector<unsigned long> > const &dims,
 	    std::vector<double const *> const &values) const;
 	bool checkParameterDim(
-	    std::vector<std::vector<unsigned int> > const &dims) const;
+	    std::vector<std::vector<unsigned long> > const &dims) const;
 	bool checkParameterValue(
 	    std::vector<double const *> const &args,
-	    std::vector<std::vector<unsigned int> > const &dims) const;
+	    std::vector<std::vector<unsigned long> > const &dims) const;
 	void evaluate(
 	    double *value,
 	    std::vector <double const *> const &args,
-	    std::vector<std::vector<unsigned int> > const &dims) const;
+	    std::vector<std::vector<unsigned long> > const &dims) const;
     };
 
 }
