+++ src/include/graph/VectorStochasticNode.h
@@ -12,8 +12,8 @@ class VectorDist;
  */
 class VectorStochasticNode : public StochasticNode {
     VectorDist const * const _dist;
-    std::vector<unsigned int> _lengths;
-    void sp(double *lower, double *upper, unsigned int length,
+    std::vector<unsigned long> _lengths;
+    void sp(double *lower, double *upper, unsigned long length,
 	    unsigned int chain) const;
 public:
     /**
@@ -31,7 +31,7 @@ public:
     bool checkParentValues(unsigned int chain) const;
     //StochasticNode *clone(std::vector<Node const *> const &parents,
     //Node const *lower, Node const *upper) const;
-    unsigned int df() const;
+    unsigned long df() const;
     double KL(unsigned int chain1, unsigned int chain2, RNG *rng,
 	      unsigned int nrep) const;
 };
